refactor: move icons to ./icons
