+++ src/assets/icons/star.png
